fixed https://github.com/BiocPy/pyBiocFileCache/issues/11
