+++ src/pybiocfilecache/const.py
@@ -1,0 +1,1 @@
+SCHEMA_VERSION = "0.99.4"
