--- tests/conftest.py
@@ -1,9 +1,0 @@
-"""Dummy conftest.py for pybiocfilecache.
-
-If you don't know what this is for, just leave it empty.
-Read more about conftest.py under:
-- https://docs.pytest.org/en/stable/fixture.html
-- https://docs.pytest.org/en/stable/writing_plugins.html
-"""
-
-# import pytest
